MR-128: modals added and logic migration to header component
